+++ .eslintrc.json
@@ -13,10 +13,7 @@
   "rules": {
     "no-shadow": "off",
     "no-param-reassign": "off",
-    "eol-last": "off",
-    "import/extensions": [ 1, {
-      "js": "always", "json": "always"
-    }]
+    "eol-last": "off"
   },
   "ignorePatterns": [
     "dist/",
