+++ date
@@ -1,0 +1,12 @@
+const today = new Date();
+const year = today.getFullYear();
+const month = today.getMonth() + 1;
+const day = today.getDate();
+
+const hour = today.getHours();
+const minutes = today.getMinutes();
+const seconds = today.getSeconds();
+
+const date = `${month}/${day}/${year}, ${hour}:${minutes}:${seconds}`;
+
+Declarations.showDate.innerHTML = `${date.toString()}`;
