+++ index.html
@@ -7,7 +7,7 @@
     <link rel="stylesheet" href="styles.css" />
     <title>Awesome Books by Talha</title>
   </head>
-  <body onload="showBooks()">
+  <body>
     <nav>
       <a href="index.html">Awesome Books</a>
 
@@ -25,7 +25,7 @@
     </nav>
 
     <div class="showDate">date</div>
-    
+
     <main>
       <div class="container">
         <div class="bookSection active">
@@ -69,7 +69,7 @@
     <footer>
       <p>Copyright Talha</p>
     </footer>
-    <script src="script.js"></script>
+    <script type="module" src="index.js"></script>
     <script src="show.js"></script>
   </body>
 </html>
