+++ modules/declarations.js
@@ -1,0 +1,15 @@
+const bookTitle = document.getElementById('title');
+const bookAuthor = document.getElementById('author');
+const addBookButton = document.getElementById('addBook');
+const booksDiv = document.getElementById('booksDiv');
+const myForm = document.querySelector('form');
+const showDate = document.querySelector('.showDate');
+
+export {
+    bookTitle,
+    bookAuthor,
+    addBookButton,
+    booksDiv,
+    myForm,
+    showDate
+};
