--- package.json
@@ -1,20 +1,0 @@
-{
-  "name": "awesome-books-es6",
-  "version": "1.0.0",
-  "description": "![](https://img.shields.io/badge/Microverse-blueviolet)",
-  "main": "script.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/tkz96/awesome-books-es6.git"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/tkz96/awesome-books-es6/issues"
-  },
-  "homepage": "https://github.com/tkz96/awesome-books-es6#readme"
-}
